Move exception handlers into their own function
